move behavior to separate file
