+++ src/resources/HorizontalPodAutoscalers/HorizontalPodAutoscalerDetails.js
@@ -13,8 +13,8 @@ import { useUrl } from 'hooks/useUrl';
 
 import { currentMetricsParser, metricsParser } from './helpers';
 import { HorizontalPodAutoscalerCreate } from './HorizontalPodAutoscalerCreate';
+import { HPABehavior } from './HPABehavior';
 import { Link } from 'react-router-dom';
-import { GenericList } from 'shared/components/GenericList/GenericList';
 
 export function HorizontalPodAutoscalerDetails(props) {
   const { t } = useTranslation();
@@ -109,91 +109,6 @@ export function HorizontalPodAutoscalerDetails(props) {
             />
           ))}
         </LayoutPanel.Body>
-      </LayoutPanel>
-    );
-  };
-
-  const HPABehavior = ({ spec }) => {
-    if (!spec?.behavior) return null;
-
-    const scaleUp = spec?.behavior?.scaleUp;
-    const scaleDown = spec?.behavior?.scaleDown;
-
-    const rowRenderer = policy => {
-      return [policy.type, policy.value, policy.periodSeconds];
-    };
-
-    return (
-      <LayoutPanel className="fd-margin--md " key="hpa-behavior">
-        <LayoutPanel.Header>
-          <LayoutPanel.Head title={t('hpas.headers.behavior')} />
-        </LayoutPanel.Header>
-        {scaleUp && (
-          <LayoutPanel className="fd-margin--md " key="hpa-behavior-scale-up">
-            <LayoutPanel.Header>
-              <LayoutPanel.Head title={t('hpas.headers.scale-up')} />
-            </LayoutPanel.Header>
-            <LayoutPanel.Body>
-              <LayoutPanelRow
-                name={t('hpas.headers.stabilization-window-seconds')}
-                value={
-                  scaleUp.stabilizationWindowSeconds ?? EMPTY_TEXT_PLACEHOLDER
-                }
-              />
-              <LayoutPanelRow
-                name={t('hpas.headers.select-policy')}
-                value={scaleUp.selectPolicy ?? EMPTY_TEXT_PLACEHOLDER}
-              />
-            </LayoutPanel.Body>
-            {scaleUp?.policies && (
-              <GenericList
-                searchSettings={{ showSearchField: false }}
-                entries={scaleUp.policies}
-                key="behavior-scale-up"
-                title={t('hpas.headers.scale-up')}
-                headerRenderer={() => [
-                  t('hpas.headers.type'),
-                  t('hpas.headers.value'),
-                  t('hpas.headers.period-seconds'),
-                ]}
-                rowRenderer={rowRenderer}
-              />
-            )}
-          </LayoutPanel>
-        )}
-        {scaleDown && (
-          <LayoutPanel className="fd-margin--md " key="hpa-behavior-scale-down">
-            <LayoutPanel.Header>
-              <LayoutPanel.Head title={t('hpas.headers.scale-down')} />
-            </LayoutPanel.Header>
-            <LayoutPanel.Body>
-              <LayoutPanelRow
-                name={t('hpas.headers.stabilization-window-seconds')}
-                value={
-                  scaleDown.stabilizationWindowSeconds ?? EMPTY_TEXT_PLACEHOLDER
-                }
-              />
-              <LayoutPanelRow
-                name={t('hpas.headers.select-policy')}
-                value={scaleDown.selectPolicy ?? EMPTY_TEXT_PLACEHOLDER}
-              />
-            </LayoutPanel.Body>
-            {scaleDown?.policies && (
-              <GenericList
-                searchSettings={{ showSearchField: false }}
-                entries={scaleDown.policies}
-                key="behavior-scale-down"
-                title={t('hpas.headers.scaleDown')}
-                headerRenderer={() => [
-                  t('hpas.headers.type'),
-                  t('hpas.headers.value'),
-                  t('hpas.headers.period-seconds'),
-                ]}
-                rowRenderer={rowRenderer}
-              />
-            )}
-          </LayoutPanel>
-        )}
       </LayoutPanel>
     );
   };
